Number of Vertical/Horizontal dividers should be an int
cf https://github.com/NREL/EnergyPlus/issues/9344#issuecomment-1077434427

+++ src/energyplus/ForwardTranslator/ForwardTranslateWindowPropertyFrameAndDivider.cpp
@@ -96,11 +96,11 @@ namespace energyplus {
     //}
 
     //if (!modelObject.isNumberOfHorizontalDividersDefaulted()){
-    idfObject.setDouble(WindowProperty_FrameAndDividerFields::NumberofHorizontalDividers, modelObject.numberOfHorizontalDividers());
+    idfObject.setInt(WindowProperty_FrameAndDividerFields::NumberofHorizontalDividers, modelObject.numberOfHorizontalDividers());
     //}
 
     //if (!modelObject.isNumberOfVerticalDividersDefaulted()){
-    idfObject.setDouble(WindowProperty_FrameAndDividerFields::NumberofVerticalDividers, modelObject.numberOfVerticalDividers());
+    idfObject.setInt(WindowProperty_FrameAndDividerFields::NumberofVerticalDividers, modelObject.numberOfVerticalDividers());
     //}
 
     //if (!modelObject.isDividerOutsideProjectionDefaulted()){
